backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography
snowflake-connector-python
httpx
python-dotenv
google-auth
apscheduler
redis
anthropic
slowapi

# Testing
pytest
pytest-asyncio
pytest-cov
